Document the limitation of using string templates in QML
Warn about using template strings in translator functions:

qsTr(`The runtime value ${val} is fetched`)

When running the application, for val=100, the qsTr method will see:

"The runtime value 100 is fetched"

And this is different from what Linguist will see and what will be
produced in the .ts files. Hence, the translation cannot be found later.

Warn user about this situation and document the reasoning behind this
limitation.

Fixes: QTBUG-127527
Change-Id: I92e517d60c47e1cdff82c8fafc5fffec60adde74
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>