Revamp documentation for Apple operating systems
The documentation is now isolated into its own directory,
and the Qt for macOS and Qt for iOS documentation has been
given a unified structure.

We now document the minimum version supported, as well
as the maximum version that we've tested, but we also
note that newer OS versions are supported, with the normal
caveats that bugs may exist, and are fixed in accordance
with the relevant branch policies.

Pick-to: 6.8 6.5
Change-Id: I99c7a445ee10d1d76b1a25223f9d0086230a0334
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
Reviewed-by: Vladimir Minenko <vladimir.minenko@qt.io>