Document how to get a CMake project to use the native photo picker
qmake automatically scans the Info.plist file to see if the project
needs to link to libqiosnsphotolibrarysupport.a, and does the linking
automatically if required. CMake currently lacks such a mechanism, so
users would need to manually link the plugin. This commit adds the
instruction on linking the plugin when using CMake.

Task-number: QTBUG-130850
Pick-to: 6.8 6.9
Change-Id: I2ddc336849140d87ef2ddc87a68c508834942472
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>