CMake: Fix CMake warnings in txtviewer plugin example
The txtviewer plugin example is part of the documentviewer example but
at the same time a separate example with its own documentation page.
That means it appears on Qt Creator's Welcome Page and must be buildable
standalone.

Add the usual CMake stanza for standalone projects to the top of the
project file.

Pick-to: 6.8 6.9
Task-number: QTBUG-132738
Change-Id: Ib5aad765ec5dbc2406df6b1cca574a20448f6529
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>