Doc: Prevent translation of CMake target names in documentation
Use the teletype (\tt) formatting command to mark strings that are
not meant to be machine-translated in the generated documentation.

Note that in section titles, this goes against our writing guidelines
as titles are not meant to have special visual formatting. Once QDoc
has a formatting command for controlling translation of strings, these
\tt commands can be replaced.

Pick-to: 6.9
Fixes: QTWEBSITE-1196
Change-Id: Ib8b92c4e75c1b334c8f0c1a52ab2c12c53ac7346
Reviewed-by: Jerome Pasion <jerome.pasion@qt.io>